Renamed page object files
